Use zinc to create the context jar for zinc and rsc (#7833)
We are not currently using `zinc`'s `-jar` option to jar our outputs, which means that in order to use the `--use-classpath-jars` option (without adding a wrapper script) we would need to capture a snapshot of loose output class files, download it, zip it, and then re-upload... which would mostly defeat the purpose of that option.

After fixing a minor issue with `zinc` (#7834), use its `-jar` option to produce the context jar. `javac_compile` continues to use the existing python codepath for this.

In order to continue to inject resources created for annotation processors and compiler plugins, even when the compiler is itself producing the output jar, we now provide the extra_resources as a `Digest` that a compiler should mix in.

`--use-classpath-jars` can be used with `hermetic` `zinc` and `rsc`.