add debug log for jvm target compile time ; fix execution graph display (#7873)
### Problem

We print out lots of relevant information to zinc/rsc compile performance with `-ldebug`, including a drawing of the execution graph, and the timings of any v2 process executions. We write per-target compile data to the run tracker which can be queried afterwards, but it would be nice to have that information written to the terminal for interactive or hacky screen-scraping use cases.

### Solution

- Write a debug log of the per-target compile data in `self._record_target_stats()`.
- Flip the direction of the `ExecutionGraph` arrows.

### Result

The debug logging during an rsc compile should hopefully be more useful!